+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.4.1-next.1...@esri/calcite-components-react@2.5.0-next.0) (2024-02-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [2.4.1-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.4.1-next.0...@esri/calcite-components-react@2.4.1-next.1) (2024-02-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [2.4.1-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.4.0...@esri/calcite-components-react@2.4.1-next.0) (2024-02-07)
 
 **Note:** Version bump only for package @esri/calcite-components-react